Redid RPS with functions
Got rid of the if statment that that would assign the computerMove variable and instead opted to make it a function that gets called. Due to scope and reasons beyond my understanding, the line that output the randomNumber to console was giving me errors, so I got rid of it.